Added playbook for software upgrade 03 testcase
