+++ ansible/regtest/files/downgrade_version
@@ -1,0 +1,3 @@
+version: b5c22f80d6b981796bdd875839325e9750beb5f0
+version: 581c0a529a332cdab0e8f767130c48fd3b8b0f01
+version: 60f39141fbbf78ddb2260dba74c68f2789374f18
