+++ ansible/regtest/files/upgrade_version
@@ -1,0 +1,3 @@
+version: 80fc5f6d01faa017caf8a0eb9726bdabca58c7c6
+version: a9d6519434efcb0163353ce84378caee96faf596
+version: 60f39141fbbf78ddb2260dba74c68f2789374f18
