+++ ansible/regtest/playbooks/upgrade/software_upgrade_03.yml
@@ -1,0 +1,99 @@
+---
+
+
+- hosts: server_emulator
+  become: true
+
+  tasks:
+    - command: "date +%Y%m%d%T"
+      register: start_time
+
+    - set_fact:
+        hash_name: "software_upgrade_03_{{ start_time.stdout }}"
+        start_time: "{{ start_time.stdout }}"
+
+
+- hosts: leaf:spine
+  become: true
+  serial: 1
+
+  vars:
+  - log_dir_path: /var/log/regression/upgrade
+
+  tasks:
+    - file:
+        path: "{{ log_dir_path }}"
+        state: directory
+
+#    - name: Download the goes installer file
+#      get_url:
+#        url: "{{ goes_installer_url }}"
+#        dest: "{{ goes_installer_dir }}{{ goes_installer_name }}"
+#        mode: 0755
+
+    - name: Copy upgrade and downgrade installer files
+      copy:
+        src: "../../files/{{ item }}"
+        dest: "{{ goes_installer_dir }}{{ item }}"
+        mode: 0755
+      with_items:
+      - "{{ goes_upgrade_installer_name }}"
+      - "{{ goes_downgrade_installer_name }}"
+
+    - name: Upgrade and downgrade goes
+      goes_upgrade_downgrade:
+        switch_name: "{{ inventory_hostname }}"
+        installer_dir: "{{ goes_installer_dir }}"
+        upgrade_installer_name: "{{ goes_upgrade_installer_name }}"
+        downgrade_installer_name: "{{ goes_downgrade_installer_name }}"
+        upgrade_version: "{{ lookup('file', '../../files/upgrade_version') }}"
+        downgrade_version: "{{ lookup('file', '../../files/downgrade_version') }}"
+        hash_name: "{{ hostvars['server_emulator']['hash_name'] }}"
+        log_dir_path: "{{ log_dir_path }}"
+      register: module_out
+
+    - command: "date +%Y%m%d%T"
+      register: end_time
+
+    - name: Fetch the log file
+      slurp:
+        src: "{{ module_out.log_file_path }}"
+      register: logs
+
+    - name: Store the test result in a hash in redis db on server emulator
+      store_result_in_redis:
+        hash_name: "{{ hostvars['server_emulator']['hash_name'] }}"
+        start_time: "{{ hostvars['server_emulator']['start_time'] }}"
+        end_time: "{{ end_time.stdout }}"
+        hash_dict: "{{ module_out.hash_dict }}"
+        log_content: "{{ logs['content'] | b64decode }}"
+      delegate_to: 127.0.0.1
+
+    - shell: "goes hget platina package | grep version:"
+      register: version_out
+
+    - shell: "goes hget platina package | grep tag:"
+      register: tag_out
+
+    - name: Store GOES version and tag details in redis db
+      store_goes_details:
+        switch_name: "{{ inventory_hostname }}"
+        hash_name: "{{ hostvars['server_emulator']['hash_name'] }}"
+        version_details: "{{ version_out.stdout }}"
+        tag_details: "{{ tag_out.stdout }}"
+      delegate_to: 127.0.0.1
+
+
+- hosts: server_emulator
+  become: true
+  tasks:
+    - name: Get the test result from redis db
+      get_test_result_from_redis:
+        hash_name: "{{ hostvars['server_emulator']['hash_name'] }}"
+      register: result
+
+    - debug:
+        msg: "Test case result: {{ result.result_status }}"
+
+    - debug:
+        msg: "Failure summary: {{ result.result_detail }}"
